Restore Windows compatibility for the exclusive-lock probe (review)
The flock-based exclusive-access check referenced syscall.Flock
unconditionally, so GOOS=windows builds failed - as they already did
before the audit cleanup: the removed runtime.GOOS branch never made
the package compile on Windows because syscall.Flock was referenced in
the same file regardless of the runtime check.

The probe now lives behind build tags: flock on Unix, LockFileEx
(non-blocking, fail-immediately) via golang.org/x/sys/windows on
Windows, so the check stays meaningful there instead of being skipped.
The flock-specific test moves to a !windows file, and CI gains a
cross-compile check for windows and darwin to catch regressions.
