+++ .github/workflows/go.yml
@@ -49,6 +49,11 @@ jobs:
       - name: Build
         run: go build -v ./...
 
+      - name: Cross-compile check (windows, darwin)
+        run: |
+          GOOS=windows go build ./...
+          GOOS=darwin go build ./...
+
       - name: Run Tests with Coverage
         run: go test -v -coverprofile=coverage.out -covermode=atomic ./...
